+++ print_lower_hex.c
@@ -6,13 +6,13 @@
 /*   By: lfuruno- <lfuruno-@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/23 15:56:51 by lfuruno-          #+#    #+#             */
-/*   Updated: 2023/11/23 16:03:14 by lfuruno-         ###   ########.fr       */
+/*   Updated: 2023/11/30 09:13:04 by lfuruno-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	print_lower_hex(unsigned long c)
+int	print_lower_hex(unsigned int c)
 {
 	char	*ptr;
 	int		count;
